[MONGO] Set UTC timezone by default (refactoring)
