+++ openframe-data-mongo-sync/src/main/java/com/openframe/data/config/NotificationContextMongoConfig.java
@@ -6,6 +6,9 @@ import com.openframe.data.repository.notification.NotificationContextWriteConver
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
+import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
+
+import java.util.List;
 
 @Configuration
 @Import(NotificationContextJacksonConfig.class)
@@ -21,6 +24,21 @@ public class NotificationContextMongoConfig {
         return new NotificationContextWriteConverter(objectMapper);
     }
 
-    // The MongoCustomConversions bean is aggregated centrally in MongoInfraConfig (which also adds the
-    // UTC LocalDate converters), so these converters are exposed as plain beans and collected there.
+    /**
+     * The {@link MongoCustomConversions} for contexts that import this config standalone (e.g. the
+     * notification integration tests, which don't bring in {@link MongoInfraConfig}). Includes the
+     * notification context converters plus the shared UTC {@link java.time.LocalDate} converters.
+     * In full application contexts this is the single conversions bean; {@code MongoInfraConfig}'s
+     * fallback is {@code @ConditionalOnMissingBean} so the two never collide.
+     */
+    @Bean
+    public MongoCustomConversions notificationContextCustomConversions(
+            NotificationContextReadConverter readConverter,
+            NotificationContextWriteConverter writeConverter) {
+        return new MongoCustomConversions(List.of(
+                readConverter,
+                writeConverter,
+                LocalDateUtcMongoConverters.LocalDateToUtcDateConverter.INSTANCE,
+                LocalDateUtcMongoConverters.UtcDateToLocalDateConverter.INSTANCE));
+    }
 }
